JedisConfig to RedisConfig
